Add JSON prototype for a newsstory
In contrast to a simple news article, a news story is a complex news article  e.g. like a Spiegel "Reportage" or an ARD "Multimedia Story". It consists of many content blocks. Each block is an article of its own and  uses all means necessary to tell the story. From audio to video everything is possible